Stop casting track numbers to int.
The track numbers for some vinyl 45 releases are not integral values but
instead have letters prepended to indicate the record side, resulting in
track numbers such as "A3", "D2", etc. This keeps those descriptions,
previously these strings would be cast to int and result in a list full of
track number '0's. The alternative is to generate track numbers based on the
index of the track within the release. I feel it's more appropriate to preserve
the information associated with the specific release. Ideally users will notice
this and choose the release for the right medium. I dunno.

Example:
http://musicbrainz.org/release/0537544d-05ad-4964-87b1-7fb15ef1bf9e