Use date and year of original release.
The release date in the generated description and the year provided via the API
now use the date of the original release.

Additionally, the year of the re-issue, if applicable, is provided in a
'release_year' field via the API. No code on site yet uses this data, as I'm
not sure whether users would be smart enough to choose the correct release, and
faulty re-issue entries would suck. But the data is there now, so we can think
about doing it if we can find something that works.